in progress work on class separation
