+++ demo/consistent-evaluation-feedback.html
@@ -7,7 +7,7 @@
 			import '@brightspace-ui/core/components/demo/demo-page.js';
 			import '@brightspace-ui/core/components/typography/styles.js';
 			import '@brightspace-ui/core/components/typography/typography.js';
-			import '../components/right-panel/consistent-evaluation-feedback-presentation.js';
+			import '../components/right-panel/consistent-evaluation-feedback.js';
 		</script>
 		<title>d2l-consistent-html-editor</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -18,13 +18,13 @@
 		<d2l-demo-page page-title="consistent-evaluation-feedback">
 
 			<d2l-demo-snippet>
-				<d2l-consistent-evaluation-feedback-presentation
+				<d2l-consistent-evaluation-feedback
 					can-edit-feedback
 					header="Feedback"
 					href=""
 					.richtextEditorConfig="{}"
 					token=""
-				></d2l-consistent-evaluation-feedback-presentation>
+				></d2l-consistent-evaluation-feedback>
 			</d2l-demo-snippet>
 
 		</d2l-demo-page>
